refactor: move audio-related functions from app to utils for better organization
